fonction delete user

+++ app/controllers/profiles_controller.rb
@@ -1,6 +1,4 @@
 class ProfilesController < ApplicationController
-  before_action :authenticate_user!, except: [:show]
-
   def show
     @user = current_user
     restrict_access unless current_user == @user
@@ -11,14 +9,21 @@ class ProfilesController < ApplicationController
   end
 
   def destroy
-    @user = current_user
-    Rails.logger.info("Trying to delete user with ID: #{@user.id}")
-    
-    if @user.destroy
-      # Ajoutez ici la logique de redirection après la suppression du profil.
-      Rails.logger.info("User deleted successfully.")
+    @user = User.find(params[:id])
+  
+    if @user
+      Rails.logger.info("Trying to delete user with ID: #{@user.id}")
+      
+      if @user.destroy
+        Rails.logger.info("User deleted successfully.")
+        render json: { message: "Compte supprimé avec succès." }, status: :ok
+      else
+        Rails.logger.error("Error deleting user: " + @user.errors.full_messages.join(', '))
+        render json: { error: "Erreur lors de la suppression du compte utilisateur." }, status: :unprocessable_entity
+      end
     else
-      Rails.logger.error("Error deleting user: " + @user.errors.full_messages.join(', '))
+      Rails.logger.error("User not found.")
+      render json: { error: "Utilisateur non trouvé." }, status: :unprocessable_entity
     end
   end
